+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed one more case (`Iterator#take`) of a V8 ~ Chromium < 126 [bug](https://issues.chromium.org/issues/336839115)
 - Forced replacement of `Iterator.{ concat, zip, zipKeyed }` in the pure version for ensuring proper wrapped `Iterator` instances as the result
 - Fixed some cases of iterators closing in `Iterator.{ zip, zipKeyed }` polyfill in the `strict` mode
+- Fixed iterator closing in `Set#{ isDisjointFrom, isSupersetOf }` polyfill
 - Fixed `DataView#setFloat16` polyfill in (0, 1) range
 - Fixed some cases of `RegExp` NCG polyfill in combination with other types of groups
 - Fixed `URLSearchParams#delete` polyfill with duplicate key-value pairs